+++ backend/de.metas.manufacturing.rest-api/src/main/java/de/metas/manufacturing/job/service/commands/ReceiveGoodsCommand.java
@@ -290,14 +290,14 @@ public class ReceiveGoodsCommand
 				final List<I_M_HU> createdLUs = HUTransformService.newInstance()
 						.tuToNewLUs(
 								tu,
-								QtyTU.ONE.toBigDecimal(),
+								QtyTU.ONE,
 								Objects.requireNonNull(newLUPIItem),
 								false);
 				lu = CollectionUtils.singleElement(createdLUs);
 			}
 			else
 			{
-				HUTransformService.newInstance().tuToExistingLU(tu, QtyTU.ONE.toBigDecimal(), lu);
+				HUTransformService.newInstance().tuToExistingLU(tu, QtyTU.ONE, lu);
 			}
 		}
